+++ packages/webpack-transpiler/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.2](https://github.com/stryker-mutator/stryker/compare/v2.0.1...v2.0.2) (2019-07-11)
+
+**Note:** Version bump only for package @stryker-mutator/webpack-transpiler
+
+
+
+
+
+## [2.0.1](https://github.com/stryker-mutator/stryker/compare/v2.0.0...v2.0.1) (2019-07-02)
+
+**Note:** Version bump only for package @stryker-mutator/webpack-transpiler
+
+
+
+
+
 # [2.0.0](https://github.com/stryker-mutator/stryker/compare/v1.3.1...v2.0.0) (2019-05-17)